+++ package.json
@@ -12,14 +12,14 @@
     "@mdi/font": "^3.6.95",
     "axios": "^0.20.0",
     "core-js": "^3.6.5",
-    "gmap-vue": "^1.4.0",
-    "jwt-decode": "^3.0.0-beta.2",
+    "gmap-vue": "^1.5.0",
+    "jwt-decode": "^3.0.0",
     "roboto-fontface": "*",
-    "vue": "^2.6.11",
+    "vue": "^2.6.12",
     "vue-axios": "^2.1.5",
     "vue-google-signin-button-directive": "^1.0.3",
-    "vue-router": "^3.2.0",
-    "vuetify": "^2.2.11",
+    "vue-router": "^3.4.8",
+    "vuetify": "^2.3.16",
     "vuex": "^3.5.1"
   },
   "devDependencies": {
@@ -30,7 +30,7 @@
     "@vue/cli-plugin-vuex": "~4.4.0",
     "@vue/cli-service": "~4.4.0",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "^1.0.3",
+    "@vue/test-utils": "^1.1.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-prettier": "^3.1.3",
@@ -38,10 +38,10 @@
     "lint-staged": "^9.5.0",
     "node-sass": "^4.12.0",
     "prettier": "^1.19.1",
-    "sass": "^1.19.0",
+    "sass": "^1.28.0",
     "sass-loader": "^8.0.2",
-    "vue-cli-plugin-vuetify": "~2.0.6",
-    "vue-template-compiler": "^2.6.11",
+    "vue-cli-plugin-vuetify": "^2.0.7",
+    "vue-template-compiler": "^2.6.12",
     "vuetify-loader": "^1.3.0"
   },
   "gitHooks": {
